refactor: split createDistillationsService into focused modules
- Extract normalizers.ts: distillation record normalization
- Extract promptBuilder.ts: prompt construction and mode guidance
- Extract snapshotBuilder.ts: job/pipeline/manual snapshot building
- Extract runDistillation.ts: orchestration logic with agent execution
- Service file reduced from 369 to 65 lines
