+++ packages/schemas/src/agent/AgentCapabilitySchema.ts
@@ -1,6 +1,5 @@
 import { z } from "zod/v4";
 import {
-  MAX_AGENT_CAPABILITIES,
   MAX_AGENT_DESCRIPTION_LENGTH,
   MAX_AGENT_NAME_LENGTH,
 } from "./AgentConstraints";
