+++ packages/schemas/src/meta/MetaSchema.ts
@@ -21,8 +21,8 @@ export function withMeta<T extends { createdAt: Date; updatedAt: Date }>(
   return { ...rest, meta: { createdAt, updatedAt } };
 }
 
-export function mapWithMeta<T extends { createdAt: Date; updatedAt: Date }>(
+export const mapWithMeta = <T extends { createdAt: Date; updatedAt: Date }>(
   records: T[],
-): (Omit<T, "createdAt" | "updatedAt"> & { meta: Meta })[] {
+): (Omit<T, "createdAt" | "updatedAt"> & { meta: Meta })[] => {
   return records.map((r) => withMeta(r));
-}
+};
